Previous Error resolved
I must not have been thinking clearly last night because I fixed last
night's error by simply including the class implementation (.cxx) file
in with the test program when I compiled. Without adding the class file
the linker would not work and threw an architecture error which of
course was not right.

09/30/2020